refactor(ui): extract diagnostics/helpers.ts — empty states, pairing view, syncAttemptsHistoryNote
Move the pure-ish helpers out of diagnostics.ts: newestPeerPing, the
four empty-state copy builders (loading, unavailable, noAttempts,
unavailableAttempts), the pairing-payload-to-command translator
pairingView (which still stashes state.pairingCommandRaw as a side
effect), and syncAttemptsHistoryNote. diagnostics.ts re-exports
syncAttemptsHistoryNote since it's consumed by other sync modules.
